Destroing of processes, scheduler modified, list iteration macro.
Started with coding the destruction of processes, had to modify
severely the scheduler code for the case that 'current' becomes
invalid due to process destruction. Needed a new process switch
without saving old state of process, which is destroyed while
running. Destroying processes that are not running is not tested
yet and will likely fail.
Introduced FORLIST, END_FORLIST for convenient list iteration.