Implemented copy on write for fork(), fixed some nasty bugs.
Copy on write functionality did not work right, with a fork() in the
userspace problem. A compiler optimization related bug was the culprit.
Formerly we had -Os, now there is no -O switch any more in the gcc options.
Therefore the size of the modules grew significantly, and the system
was at first broken. By some minor adjustments, also in the loader
scripts for the kernel (better alignment) this could be all resolved.