fork() is now right with COW, invalidate page info is done in map_page.
First a mistake in fork() was corrected: For COW not only the user pages
of the new process must be set read-only, also those of the original
process. Also it is absolutely necessary to update the page table
information after a new mapping is introduced in the page table system.
I do it with set_cr3(get_cr3()) because the i386 has not yet the
invlpg instruction. This page table issue caused a nasty bug in fork()
which appeared only on real hardware, not on bochs.