Time functions, usermode prog loaded from ext2 filesystem, cross-compiled.
Now seconds are counted in the timer interrupt, simple mktime and gmtime
facilities in klib, RTC is read at the start.
Ext2 read is used to load code of primary user process from the file
/user.bin in the filesystem. The process, which does this then becomes
the first user mode process. So there is no need any longer to embed the
user process in the kernel. The makefiles and loader scripts are rewritten
accordingly.
Also now building with an i686-elf-gcc cross-compiler.